requirements.txt
openai~=0.27.2
numpy~=1.23.5
PyAudio~=0.2.13
pynacl
requests~=2.28.2
typing~=3.7.4.3
discord.py @ git+https://github.com/Sheepposu/discord.py@voice-receive
SpeechRecognition
Pillow
tiktoken==0.3.3
scipy
requests